+++ lib/persistentqueue/fastqueue.go
@@ -2,11 +2,13 @@ package persistentqueue
 
 import (
 	"fmt"
+	"math"
 	"path/filepath"
 	"sync"
 
 	"github.com/VictoriaMetrics/VictoriaMetrics/lib/bytesutil"
 	"github.com/VictoriaMetrics/VictoriaMetrics/lib/fasttime"
+	"github.com/VictoriaMetrics/VictoriaMetrics/lib/fs"
 	"github.com/VictoriaMetrics/VictoriaMetrics/lib/logger"
 	"github.com/VictoriaMetrics/metrics"
 )
@@ -61,6 +63,15 @@ func MustOpenFastQueue(path, name string, maxInmemoryBlocks int, maxPendingBytes
 		n := fq.pq.GetPendingBytes()
 		fq.mu.Unlock()
 		return float64(n)
+	})
+
+	_ = metrics.GetOrCreateGauge(fmt.Sprintf(`vm_persistentqueue_free_disk_space_bytes{path=%q}`, path), func() float64 {
+		freeSpaceBytes := fs.MustGetFreeSpace(path)
+		// Limited by disk space if remoteWrite.maxDiskUsagePerURL wasn't set
+		if maxPendingBytes == 0 {
+			return float64(freeSpaceBytes)
+		}
+		return math.Min(float64(maxPendingBytes), float64(freeSpaceBytes))
 	})
 
 	pendingBytes := fq.GetPendingBytes()